Retrigger MVSQ-filtered property when set to `null`
* Retrigger MVSQ-filtered property when set to `null`

Previously, when a property was only used in a filter/return, the MVSQ compiler would specify a constraint on it with type `Any`. However, `Any` only matches when the value is set or changes between set values. When the value is null, `Any` does not match. `Unconditional` is added to match any time `Any` would match, as well as when the value is null/unset (i.e., the same time that `None` would match).

This allows where clauses in MVSQ like the following to trigger as expected:

`WHERE x.tagged IS NULL OR x.tagged = false`

Previously, this query would only trigger when x was set to false.

* extend list

* Scalafmt

* Fix test

* Fix test 2

GitOrigin-RevId: 415e1b27c6fbac56a91cf3d97fc3f11368bbc249