Add a "cassandra" option to Novelty
* Add a "cassandra" option to Novelty

Currently the root slice gets a keyspace called "novelty" (by default),
and others get a keyspace called "novelty_<context name>".
Cassandra keyspace names are limited to alphanumeric and underscore
characters, and also limited to 48 characters. With the "novelty_"
prefix, that that leaves 40 characters for the context name.

We might wish to hash the context name or something to deal with this
more generally (e.g. and then use that number as the keyspace name).
(MapDB / RocksDB impls escape the $ character from context name so as to
not conflict with the root partition, named $.)

* Address Future[Unit] discard (Unit expected)

* remove inscrutable "TODO"

GitOrigin-RevId: e257183162a628c5a056bd47b1ca8aeb48c2fb87