SSL option for Quine OSS, Enterprise, and Novelty
* Switch from deprecated java.net.URL constructor to Akka's Uri

The public constructor for java.net.URL is deprecated. Akka's Uri
class works, and has Scala-friendly methods, and generally seems nicer
to work with.

* Determine resolvable URL once, pass that into the Recipe for status
query output. Previously this feature required setting an advertised URL
in the settings, which I'm assuming was an oversight. It now defaults to
a URL based on the bind address if an advertised URL is not set.

* Use helper functions for bind address stuff

Novelty just does its own thing as far as this bind address stuff.

GitOrigin-RevId: b2a1b36c0c4d3f2bce5015a555de9f240b589971