Keep graph in ready state until shutdown completes
* Keep graph in ready state until shutdown completes

The graph becoming unready as soon as shutdown is initiated prevents
CypherQuery standing query outputs from being processed. These last
results will continue retrying until the "before-cluster-shutdown"
phase times out, abruptly terminating the output stream, causing
warnings. This change keeps the graph in a ready state while it is
shutting down, allowing the final outputs to be processed.

QU-1521

* Change an ExecutionContext

We should take a look at the surrounding uses of shardDispatcherEC, but
this at least makes the additional cheap code run on the calling thread.

QU-1521

* Change surrounding ExecutionContexts to nodeDispatcherEC

shardDispatchEC was previously an implicit that was picked up in many unintended
call sites. It was made explicit to allow incrementally re-evaluating where it is used.

The documentation for the execution contexts (node, shard, blocking) in reference.conf
says this about the shardDispatcherER:
"This should rarely, if ever, be used for scheduling raw tasks like `Future`s."

The general purpose nodeDispatcherEC seems more appropriate for shutdown here and in
the base class, so they have been changed.

QU-1521

GitOrigin-RevId: e4cd4a596b3743396e3c2e2b559a90d30a2d2e47