Make a "newtype wrapper" NamespaceParameter.
* Make a "newtype wrapper" NamespaceParameter.

Now it's an actual type (as opposed to a type alias), and thus can have
unique typeclass instances. The prior instances applied to all things
that type alias normalized to: Option[String], which caused issues where
that type was used elsewhere.

This type is a "value class" wrapper around String - and theoretically
should have zero extra allocation cost over String by having the same
runtime representation as a String - though that efficiency doesn't
really matter for this case.

The main point is the class has a private constructor, and an apply
method on the companion object for creating valid instances of
`NamespaceParameter` - so that all instances are guaranteed to be all
lowercase and to match the character rules.

* Removed the check for namespace existence and throw from the `Codec`
instance for `NamespaceParam`. The `override def handleServerError` can
be removed from `BaseAppRoutes`, too.

* Moved `namespaceFromParam` from the quine-core project's `graph`
package object to the quine project's `ServerQuineEndpoints` - as
quine-core isn't aware of quine-endpoint's `NamespaceParamter` type,
while quine is.

* Remove `newNamespacePathParameter` in favor of `namespacePathParameter`

* Finish replacing "String" with "NamespaceParameter" on the create namespace route

* Clean up the Codec situation

Define `namespaceCodec` as `Codec[String, NamespaceParameter]`, use that
for the `Segment` and `JsonSchema` instances.

Define `optionalNamespaceCodec: Codec[Option[String],
NamespaceParameter]` in terms of that (by wrapping it), and use that for
the `NamespaceParameter` instance.

Endpoints4s as a `Coded.sequentially` method for combining codecs, if we
wanted to define an `Codec[Option[String], String]` to combine with the
`Codec[String, NamespaceParameter]` to do the defaulting for us, instead
of the "wrapping" delegating to the first Codec. But that just felt
weird to me.

---------

Co-authored-by: Leif Warner <leif@Leifs-MacBook-Pro.local>
GitOrigin-RevId: 8bdb369c1e7234a7393ee7c9eb85c4a81623d054