Add ExecutionContexts as a parameter to MapDb.
* Add ExecutionContexts as a parameter to MapDb.

For use with ScalaTest's SerialExecutionContext - to make test order
more deterministic.

Defined a new `ComputeAndBlockingExecutionContext` interface to pass in
to MapDb, made QuineDistpacthes a subclass of that, and then added a
test-only `FromSingleExecutionContext` impl that uses the same
ExecutionContext for both.

Used in MapDbPersistorSpec.
MapDbPersistorTests times out if I try to use SerialExecutionContext -
maybe it's blocking some threads or otherwise relies on multithreading?

* Use SerialExecutionContext in RocksDbPersistorSpec too

* Switched MapDb and RocksDb to use parasitic ExecutionContext in tests

https://github.com/scalatest/scalatest/issues/2314#issuecomment-1978816544

* Also convert HistoricalQueryTests to use parasitic ExecutionContext

GitOrigin-RevId: d83efbcd7126be0a2ad9f6294a03ab52a5e1c54c