QU-1793: Websocket connection to LSP
# Motivation

This PR creates a websocket connection exposing the
`QuineLanguageServer` object which interfaces with `lsp4j`.

# Implementation

This PR creates the websocket endpoint at `api/v1/lsp` which currently
only accepts a TextMessage of `Initialize`. The websocket message stream
flows into an Actor (of which there is 1 Actor per websocket connection)
which is responsible for containing an instance of `QuineLanguageServer`
that the client manipulates through messages to the endpoint.

There are 3 main files in this PR:

- `WebSocketQuinePatternServer.scala`: Contains the websocket endpoint,
the messages stream handler, and instantiates an Actor for the
connection
- `QuineLanguageServerActor.scala`: This actor is responsible for
holding an instance of `QuineLanguageServer`, and
  1. accepts a stream of `Message`s
  2. parses them into a `ClientMessage`
3. does appropriate logic based on the `ClientMessage` (right now, only
`Initialize` is handled)
  4. emits a `ServerMessage`, parsed into a `TextMessage`
- `WebSocketQuinePatternServerTest.scala`: This tests the messages flow
handler, ensuring expected responses.

---

## Testing externally

In the picture below, I open up a websocket connection, subscribe to the
messages returned from the server, and then send JSON stringified
strings (later, these will be stringified objects). The `Initialize`
message correctly returns an `InitializeSuccess` message, and other text
messages correctly returns `UnsupportedTextMessage`.

![image](https://github.com/thatdot/quine-plus/assets/13839150/90976cda-4d65-481d-9813-3663b055fa39)

---

https://github.com/thatdot/quine-plus/assets/13839150/ec2292a9-f71b-49df-a924-4849e7e63b33

GitOrigin-RevId: a2aed2c52f82fa09626da0ad07f18f26556c6d21