Remove explicit dep on kotlin
# Description

The comment said it was to use a more recent version to get past a CVE
from the version depended on via MapDB. That was likely obsoleted via
the MapDB update https://github.com/thatdot/quine-plus/pull/2702
which effectively changed the kotlin dep to "the latest 1.9.x", meaning
we were already on the latest, and the value we specified here for
kotlin-stdlib was irrelevant.
This value was also being used for kotlin-stdlib-jdk8, but the comment
didn't mention that, and that's not a dep of mapdb.

## Type of change

- [x] Refactor (rebased to a single commit to keep separate from other
changes)
# How Has This Been Tested?
Evaluated the results of sbt tasks `whatDependsOn org.jetbrains.kotlin
kotlin-stdlib` and `whatDependsOn org.jetbrains.kotlin
kotlin-stdlib-jdk8`

# Checklist:
- [x] I have performed a self-review of my code
- [x] My changes generate no new warnings
- [x] New and existing unit tests pass locally with my changes

Closes https://github.com/thatdot/quine-plus/pull/2835

GitOrigin-RevId: 9ba4b57f106baabdbeff528f82657bd4012fcc28