Remove direct dep on snakeyaml.
# Description
This same version is already a transitive dep of circe-yaml, which we're
using. We do use a snakeyaml class directly in
ServerEntitiesWithExamples.yamlRequest, but that's simply the line
`Parser.default.decodeAccumulating(new
YamlUnicodeReader(requestInputStream))`
`Parser.default.decodeAccumulating` is from circe-yaml, and
`YamlUnicodeReader` is a util class from snakeyaml-engine implementing a
java.io.Reader (mainly using to ensure proper UTF8 handling). While we
are using it directly, it's just one tiny use, and directly in service
of circe-yaml, which already includes it. Doensn't seem worth wondering
why we're explicitly specifying a dep on it.

Also refined comment as to why why have a dep on pekk-http-xml

## Type of change
Delete options that are not relevant, add if necessary
- [x] Refactor (rebased to a single commit to keep separate from other
changes)
# How Has This Been Tested?
Evaluated output of `whatDependsOn org.snakeyaml snakeyaml-engine`
# Checklist:
- [x] I have performed a self-review of my code
- [x] My changes generate no new warnings
- [x] New and existing unit tests pass locally with my changes

GitOrigin-RevId: 3d1805b1f33793c6ef26a18aebe2e9b1a678ce44