QU-1816/QU-1774: Use self-hosted runner for long CI worflows
# Description
This PR improves CI time and cost by transitioning from github-hosted
runners to self-hosted runners on AWS. In the product-e2e-testing
repository, we had a lot of success by using a small github action to
start up a larger, self-hosted runner on AWS, so this PR is migrating
this repository to using the same technique.

This PR also incidentally addresses QU-1774 by switching the pekko
artery port to a random available port for test suites, since that issue
was exacerbated by the larger machine.

QU-1816

QU-1774

## Type of change

- [x] Refactor (rebased to a single commit to keep separate from other
changes)

# How Has This Been Tested?

Describe the tests performed, including instructions to reproduce, and
relevant details for your test configuration
- [x] Ran updated Build and Test Workflow manually:
https://github.com/thatdot/quine-plus/actions/runs/8493515099
- [x] Ran updated Development Release Workflow manually:
https://github.com/thatdot/quine-plus/actions/runs/8493515393
- [x] Ran updated Integration Test Workflow manually:
https://github.com/thatdot/quine-plus/actions/runs/8493517013

# Checklist:

- [x] I have performed a self-review of my code
- [x] I have verified my code doesn't add an implementation for
something that already exists
- [x] I have commented my code, particularly in hard-to-understand areas
- [ ] ~~I have made corresponding changes to the documentation~~
- [x] My changes generate no new warnings
- [ ] ~~I have added tests that prove my change is effective and works~~
- [x] New and existing unit tests pass locally with my changes
- [x] Any changes required of upstream modules have been updated and
have explicit tests in this PR
- [x] Any downstream modules depending on this change have an
integration plan described in the description

GitOrigin-RevId: e2c9a3347285e81848c60ad4b628563034fa46fb