Property Set Fixes
# Description
Fixes a few issues in `SetProperty` and `SetProperties`:
- setting a property to `null` is no longer rewritten by the openCypher
compiler as a `RemovePropertyItem`, so reintroduce the rewrite at the
Quine interpreter level. At this level, we are also able to enforce the
semantics even when static analysis doesn't catch that a value is
nullable.
- setting multiple properties via sequential `SET` statements now
produces a `SetProperties` instance that _adds to_ rather than
_replacing_ the node's properties
- All places that construct a `PropertyRemoved` event, including
`SetProperty`/`SetProperties`, now use the correct value for the
`PropertyRemoved` value field (i.e., the previous value of the property)
- `SetProperties` is now marked idempotent, for consistency with
`SetProperty`

This PR exacerbates the problem of misusing `CypherExpressions.scala`
for tests that are not about expressions -- I will open up a follow-up
PR to move the misplaced tests to more appropriate locations. I chose
not to do it in this PR to keep the diff more legible.

## Type of change
Delete options that are not relevant, add if necessary
- [x] Bug fix (non-breaking change which fixes an issue)

# How Has This Been Tested?
Describe the tests performed, including instructions to reproduce, and
relevant details for your test configuration
- [x] Unit tests were added covering all combinations of property-based
`SET` clauses
- [x] The query we wrote to support an OSS user has the expected results

# Checklist:
- [x] I have performed a self-review of my code
- [x] I have verified my code doesn't add an implementation for
something that already exists
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have added tests that prove my change is effective and works
- [x] New and existing unit tests pass locally with my changes

GitOrigin-RevId: 2a7eb02e06154dea55f2478e18f719a33f841d4e