thatDot Streaming Graph: Add PEM option for ClickHouse cert
# Description
Adds the ability to configure a raw PEM file as a trusted certificate
for connecting to ClickHouse, as an alternative to the conventional (for
java) truststore.

This depends on a new library, "sslcontext-kickstart", from
https://github.com/Hakky54/sslcontext-kickstart. This servers a similar
role to the [SSLContextFactory shipped with post-pekko versions of akka
http](https://doc.akka.io/api/akka-http/10.6.2/akka/http/scaladsl/common/SSLContextFactory$.html).

## Type of change
Delete options that are not relevant, add if necessary
- [x] New feature (non-breaking change which adds functionality)
- [x] This change requires a documentation update

# How Has This Been Tested?
Describe the tests performed, including instructions to reproduce, and
relevant details for your test configuration
- [x] Unit test added to ClickHouseSpec to validate a connection via a
PEM file (similar to the test that validates a connection via a
truststore)
- [x] ClickHouse deployed to a local test environment, configured to
connect with a PEM file, and ethereum recipe executed with shard sizes
set to zero (to force persistence activity)

# Checklist:
- [x] I have performed a self-review of my code
- [x] I have verified my code doesn't add an implementation for
something that already exists
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my change is effective and works
- [x] New and existing unit tests pass locally with my changes

GitOrigin-RevId: 341e08c988479954131a4030b60f369300789871