Protobuf Decode Cypher procedure
# Description
Adds a procedure for decoding a protobuf value from within Cypher, given
the byte string, a schema, and a type name. Also fixes a few bugs in the
extant Protobuf tests and adds throwable documentation to the protobuf
utilities.

In order to avoid performing a file read on every invocation of
`parseProtobuf` (name chosen to match `parseJson`), the `ProtobufParser`
has been wrapped with a cache, owned by the AppState. This cache will
retain up to 10 schema-type pairs, cycling through them as necessary.
This slightly aids the protobuf construction problem by making it more
explicit, putting the blocking work on the blocking dispatcher.

## Type of change
Delete options that are not relevant, add if necessary
- [x] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [x] This change requires a documentation update

# How Has This Been Tested?
Describe the tests performed, including instructions to reproduce, and
relevant details for your test configuration
- [x] New unit tests added and pass

# Checklist:
- [x] I have performed a self-review of my code
- [x] I have verified my code doesn't add an implementation for
something that already exists
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my change is effective and works
- [x] New and existing unit tests pass locally with my changes

GitOrigin-RevId: 1100243f70feb9dba91ac7bdd50ed0b2a6f6eda5