Increase ProtobufParser cache size
# Description
Bumps the ProtobufParser cache size to 800 records. One of our customers
has a very large number of protobuf types, so a cache of 10 type
handlers will not be sufficient.

## Type of change
- [x] New feature (non-breaking change which adds functionality)

# How Has This Been Tested?
Describe the tests performed, including instructions to reproduce, and
relevant details for your test configuration
- [x] Confirmed via intellij's memory agent/parser the typical size of a
ProtobufParser to ensure it wouldn't swamp our memory
- [x] swapped the ProtobufParser unit test cache for a LoadingCache and
verified that the test still passes

# Checklist:
- [x] I have performed a self-review of my code
- [x] I have verified my code doesn't add an implementation for
something that already exists
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] My changes generate no new warnings
- [x] New and existing unit tests pass locally with my changes

GitOrigin-RevId: b6f1d702e200a9367e415a92d05964adb35d2a5e