QU-776: Disabled codes when no progress bar exists
# Description
Disabled control codes when no progress bar exists so they do not
clutter the console output.

## Type of change
- [X] Bug fix (non-breaking change which fixes an issue)
# How Has This Been Tested?
Manual test: Run a recipe to make sure that this did not mess up
progress bars and ran a case where there was no progress bar and piped
the error output to a file. Previously this file would contain the
control codes but now it does not.
To reproduce error call `sbt quine/run &> output.txt` and inspect the
output.txt with a file editor.
# Checklist:
- [X] I have performed a self-review of my code
- [X] I have verified my code doesn't add an implementation for
something that already exists
- [X] I have commented my code, particularly in hard-to-understand areas
- [ ] ~~I have made corresponding changes to the documentation~~
- [X] My changes generate no new warnings
- [ ] ~~I have added tests that prove my change is effective and works~~
- [X] New and existing unit tests pass locally with my changes
- [X] Any changes required of upstream modules have been updated and
have explicit tests in this PR
- [X] Any downstream modules depending on this change have an
integration plan described in the description

---------

Co-authored-by: David Masters <david@thatdot.com>
GitOrigin-RevId: f254806034277a1ac6e96d418e10233205b64ea1