Resolve short-name-ambiguous protobuf schemas by full name
# Description
Adds support for fully-qualified message type names to protobuf schema
interfaces. Previously, only the short name could be used. When the same
short name was reused by multiple proto files compiled into the same
desc file, a "no message of type" error was thrown.

This reimplements the protobuf schema resolution logic to use the
first-party DynamicSchema (which, as far as I can tell and remember,
wasn't used at the time simply because we didn't find it -- the docs for
this library are horrible). This means we no longer need to roll our own
dependency resolution logic.

Finally, note that resolution logic has been moved to the companion
object for protobuf schema, as it relies on effects that we'd ideally
like to avoid resolving in the constructor. This will also make QU-1868
that much easier (i.e, rewriting the schema cache in terms of
DynamicSchemas rather than Descriptors or the current file-typename
pairs).

Best of all, by coincidence, this is fully backwards-compatible!

## Type of change
Delete options that are not relevant, add if necessary
- [x] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)

# How Has This Been Tested?
Describe the tests performed, including instructions to reproduce, and
relevant details for your test configuration
- [x] Added a new protobuf test that uses a descriptor built from
multiple proto files, including at least one type that has an ambiguous
short name but an umambiguous fully-qualified name. The full name should
succeed to parse, while the short name should fail with an
"ambiguous"-type exception
- [x] Existing protobuf unit tests pass

# Checklist:
- [x] I have performed a self-review of my code
- [x] I have verified my code doesn't add an implementation for
something that already exists
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my change is effective and works
- [x] New and existing unit tests pass locally with my changes

GitOrigin-RevId: ced19e9e621a44913bb35dd2f9b8a3bc36c07c24