Bring write-side protobuf code up to parity with read-side
# Description
The overall goal of this PR is to bring the Quine->Protobuf conversions
up to parity with the recent work done on the Protobuf->Quine side. In
paricular, this PR:
- Refactors the protobuf schema cache to store DynamicSchemas (the
parsed form of `desc` files) instead of ProtobufParsers
- Reuses the appstate-owned schema cache for both reading and writing
- Adds a toProtobuf Cypher procedure mirroring the parseProtobuf (naming
convention following the JSON functions, as before)
- Lifts the side-effects (file IO) necessary to use a protobuf schema
higher up the stack, folding it in to existing Futures where possible,
and documenting where it is not possible.

## Type of change
Delete options that are not relevant, add if necessary
- [x] New feature (non-breaking change which adds functionality)
- [x] Refactor (rebased to a single commit to keep separate from other
changes)

# How Has This Been Tested?
Describe the tests performed, including instructions to reproduce, and
relevant details for your test configuration
- [x] New unit test covering `toProtobuf` procedure passes
- [x] old unit tests pass

# Checklist:
- [x] I have performed a self-review of my code
- [x] I have verified my code doesn't add an implementation for
something that already exists
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my change is effective and works
- [x] New and existing unit tests pass locally with my changes

GitOrigin-RevId: 255c6ae7b96b64deca4a916a1172ba4d343497db