Add note that we now really do depend on serde
While implementing #2915, I mistook `DynamicSchema` as coming from the
com.google.protobuf package. In actuality, it comes from the `apicurio`
package within `schema-registry-serde`. This calls that dependency out
explicitly in our build.sbt to ensure we keep the direct dependency, as
we now use it directly (before, it was only used transitively).

GitOrigin-RevId: 9d0e8af20af24643f03de3980caa4c2114802b6e