QU-1841: Update versions for common GitHub actions
# Description

This PR updates versions for common "official" (from a trusted source)
GitHub actions. This is to address warnings as a result of GitHub
deprecating node versions 12 and 16 from GitHub actions. Once these
versions are deprecated, actions which rely on node will run on version
20 regardless of which version is specified by the action.

## Type of change

Delete options that are not relevant, add if necessary
- [x] Bug fix (non-breaking change which fixes an issue)

# How Has This Been Tested?
Describe the tests performed, including instructions to reproduce, and
relevant details for your test configuration
- [ ] CI Passes for this PR

# Checklist:
- [x] I have performed a self-review of my code
- [x] I have verified my code doesn't add an implementation for
something that already exists
- [ ] ~~I have commented my code, particularly in hard-to-understand
areas~~
- [ ] ~~I have made corresponding changes to the documentation~~
- [x] My changes generate no new warnings
- [ ] ~~I have added tests that prove my change is effective and works~~
- [x] New and existing unit tests pass locally with my changes
- [x] Any changes required of upstream modules have been updated and
have explicit tests in this PR
- [x] Any downstream modules depending on this change have an
integration plan described in the description

GitOrigin-RevId: 4a43c8f7d520e1c78d9d032be3769e931d61ee21