Fix: Remove explicit dependency on aws-java-sdk-sts
# Description

At one point, we needed this explicit dependency to avoid pulling in a
vulnerable transitive dependency through `schema-registry-serde`. Now,
`schema-registry-serde` depends on a new enough version of
`aws-java-sdk-sts` to not be vulnerable, making the explicit dependency
override unnecessary.

Removing it was originally recommended by Leif in this comment:
https://github.com/thatdot/quine-plus/pull/2823#issuecomment-2032740778

I agree with the suggestion because we can avoid the extra work of
tracking an additional explicit dependency.

## Type of change
Delete options that are not relevant, add if necessary
- [x] Refactor (rebased to a single commit to keep separate from other
changes)

# How Has This Been Tested?
Describe the tests performed, including instructions to reproduce, and
relevant details for your test configuration
- [x] Dependency Track Scan

# Checklist:
- [x] I have performed a self-review of my code
- [x] I have verified my code doesn't add an implementation for
something that already exists
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] ~~I have added tests that prove my change is effective and works~~
- [x] New and existing unit tests pass locally with my changes
- [x] Any changes required of upstream modules have been updated and
have explicit tests in this PR
- [x] Any downstream modules depending on this change have an
integration plan described in the description

GitOrigin-RevId: dcce3aaa53eb68c59e66d1045f54a32031208e09