QU-1841: Update tibdex/github-app-token action to latest release
# Description
The tibdex/github-app-token action is a 3rd party developed action so it
is versioned to an exact git commit. This PR bumps the configured commit
to the latest release of the action:
https://github.com/tibdex/github-app-token/releases/tag/v2.1.0

The older version gets a warning from GitHub about using version 16 of
node which is deprecated in GitHub actions.

Incidentally, this PR also adds a name to the copybara action so it
shows up pretty in the actions tab.

## Type of change
Delete options that are not relevant, add if necessary
- [x] Bug fix (non-breaking change which fixes an issue)

# How Has This Been Tested?
Describe the tests performed, including instructions to reproduce, and
relevant details for your test configuration
Not tested.

# Checklist:
- [x] I have performed a self-review of my code
- [x] I have verified my code doesn't add an implementation for
something that already exists
- [ ] ~~I have commented my code, particularly in hard-to-understand
areas~~
- [ ] ~~I have made corresponding changes to the documentation~~
- [x] My changes generate no new warnings
- [ ] ~~I have added tests that prove my change is effective and works~~
- [x] New and existing unit tests pass locally with my changes
- [x] Any changes required of upstream modules have been updated and
have explicit tests in this PR
- [x] Any downstream modules depending on this change have an
integration plan described in the description

GitOrigin-RevId: 58f621c9a1616cca808a380acfe919c57fa30cfa