Fix: Relax RawValuesIngestSrcDef unit test timing constraints
# Description
After experiencing two different spurious failures of the
RawValuesIngestSrcDefTest, I believe it is due to an inexact timing
requirement on ingest throttles. Occasionally it appears the ingest will
only take 3 records during a 1 second period with a maxPerSecond of 4.
In this PR I have addressed the issue by relaxing the expected number of
records to be 4 +/- 1.

-
https://github.com/thatdot/quine-plus/actions/runs/9085997687/job/24970648694
-
https://github.com/thatdot/quine-plus/actions/runs/9086457362/job/24972177960

I'm putting this PR up to help, though happy to consider a different
solution if anyone has a better idea for how to address this.

## Type of change
Delete options that are not relevant, add if necessary
- [x] Bug fix (non-breaking change which fixes an issue)

# How Has This Been Tested?
Describe the tests performed, including instructions to reproduce, and
relevant details for your test configuration
- [x] CI Passes on this PR

# Checklist:
- [x] I have performed a self-review of my code
- [x] I have verified my code doesn't add an implementation for
something that already exists
- [ ] ~~I have commented my code, particularly in hard-to-understand
areas~~
- [ ] ~~I have made corresponding changes to the documentation~~
- [x] My changes generate no new warnings
- [x] I have added tests that prove my change is effective and works
- [x] New and existing unit tests pass locally with my changes
- [x] Any changes required of upstream modules have been updated and
have explicit tests in this PR
- [x] Any downstream modules depending on this change have an
integration plan described in the description

GitOrigin-RevId: f3966c40b8e93d2ef2400a3b806492ad2b35e68f