remove redundant failing test
# Description
A redundant test on Ingests is failing on the updated CI for timing
reasons. Since it's functionality is duplicated by other cases in the
same test, it's been removed.

QU-####
## Type of change
Delete options that are not relevant, add if necessary
- [X] Bug fix (non-breaking change which fixes an issue)
# How Has This Been Tested?
Describe the tests performed, including instructions to reproduce, and
relevant details for your test configuration
- [X] Unit tests
# Checklist:
- [X] I have performed a self-review of my code
- [X] I have verified my code doesn't add an implementation for
something that already exists
- [X] My changes generate no new warnings
- [X] New and existing unit tests pass locally with my changes

GitOrigin-RevId: a2820defb9325fbf398e56432d228a7ec7cb511d