QU-1832: Cleanup some old changes: Remove unnecessary synchronized block, Remove IO from IngestSrcDef
# Description

There were some "warts" introduced by a group of older PRs. This PR is
meant to address those issues with the benefit of hindsight.

Changes:

- Remove an unnecessary `synchronized` block from a function which does
not add any safety in the best case and could potentially cause a
deadlock in the worst case.
- Remove usage of cats.effect.IO in favor of `() => Future[A]` style of
variables required to expose the controls of a restartable ingest. This
effected primarily `IngestSrcDef` and its usage.

Note: The dependency on cats.effect.IO remains due to use of IO in
[PostgresDbForAWSReporter](https://github.com/thatdot/quine-plus/blob/d6e67e6e7d94420cb9c6ed749de7c972d897bb4a/novelty-base/src/main/scala/com/thatdot/novelty/app/reporters/PostgresDbForAWSReporter.scala).
I felt that it was outside of the scope of this issue/PR to refactor
that.

## Type of change
- [x] Refactor (rebased to a single commit to keep separate from other
changes)

# How Has This Been Tested?
Describe the tests performed, including instructions to reproduce, and
relevant details for your test configuration
- [x] Manually Test ingest status
- [x] Run the end to end tests

# Checklist:
- [x] I have performed a self-review of my code
- [x] I have verified my code doesn't add an implementation for
something that already exists
- [x] I have commented my code, particularly in hard-to-understand areas
- [ ] ~~I have made corresponding changes to the documentation~~
- [x] My changes generate no new warnings
- [ ] ~~I have added tests that prove my change is effective and works~~
- [x] New and existing unit tests pass locally with my changes
- [x] Any changes required of upstream modules have been updated and
have explicit tests in this PR
- [x] Any downstream modules depending on this change have an
integration plan described in the description

GitOrigin-RevId: 3000e84b6d7697b41f079433df85ea2326c2d281