QU-1917: Rebrand thatDot Streaming Graph
# Description
Finish rebranding `quine-enterprise`. There are a few things that were
not rebranded due to `quine` and `quine-enterprise` sharing the same UI:

- [x] Browser Page Title
- [x] Favicon
- [x] Logo in Exploration UI
- [x] Remove white background and make BG transparent for Quine Favicon

This PR allows `quine-enterprise` and `quine` to indicate to the React
app powering the exploration UI which product is launching the UI,
allowing the UI to conditionally render depending on which product the
user launched.

Favicons are generated using: https://realfavicongenerator.net/

## Type of change
Delete options that are not relevant, add if necessary
- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Refactor (rebased to a single commit to keep separate from other
changes)
- [ ] This change requires a documentation update
# How Has This Been Tested?
Describe the tests performed, including instructions to reproduce, and
relevant details for your test configuration
- [x] Ran `sbt quine/run` and visually inspected the Exploration UI and
the Browser Page Tab (Favicon and Title)
- [x] Did the same thing as above, but with `sbt "quine-enterprise/run"`
# Checklist:
- [x] I have performed a self-review of my code
- [x] I have verified my code doesn't add an implementation for
something that already exists
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] My changes generate no new warnings
- [x] New and existing unit tests pass locally with my changes
- [x] Any changes required of upstream modules have been updated and
have explicit tests in this PR
- [x] Any downstream modules depending on this change have an
integration plan described in the description

---

# Quine Favicon, Title, and TopBar Logo

![image](https://github.com/thatdot/quine-plus/assets/13839150/2f0919b5-bffc-40ee-b3b7-b0145f14bc5f)

# thatDot Streaming Graph Favicon, Title, and TopBar Logo

![image](https://github.com/thatdot/quine-plus/assets/13839150/6d8d863f-6bcf-4cd7-8c20-abba7448fa8d)

GitOrigin-RevId: 4ffd0df0600510f776383e5e21205abe15fb28db