Add migration infrastructure and a "rewrote MVSQ" migration
# Description
Introduces a migration system to Quine and thatDot Streaming Graph to
perform out-of-band migration operations on behalf of the user, or to
give the user instructions on manual steps they need to perform to
upgrade their product version.

This is accomplished by introducing:
- `MigrationVersion`, a type for managing a 4-byte persisted (in
metadata) value to identify the latest migration version successfully
applied by the product
- `Migration`, a type for representing each breaking change that
requires some migration effort. Migrations have a `from` and a `to`
`MigrationVersion`, and are expected to be sequential and contiguous in
the registry at `migrations.instances.all`
- `Migration.Apply`, a typeclass whose instances describe how to apply
any given `Migration`. For example, there are 2
`Migration.Apply[MultipleValuesRewrite]` instances. One which runs in
Quine, and knows how to advise a Quine user how to migrate or drop their
persistor, and one which runs in thatDot Streaming Graph, and knows how
to advise a thatDot Streaming Graph user how to migrate or drop their
persistor. The typeclass pattern was chosen because it enables us to
include application-specific checks (eg, quine-enterprise checking for
ClickHouse) in the implementations
- Logic at application load time to check the last-persisted
MigrationVersion and, if it is outdated, apply any not-yet-applied
migrations

The intended pattern for using this system is, for any breaking change,
add a `Migration` (and "register" it to `migrations.instances.all` by
adding it to that `Seq`) and add a `Migration.Apply` instance for that
migration for each product. If the migration should be applied
identically regardless of product, the same `Migration.Apply` instance
can be reused.

This PR does not complete the implementation of the migration system,
and the following corners are cut:
- There is no dynamic resolution of `Migration.Apply` instances -- right
now, with only one migration in each product, those instances are simply
hard-coded with a `require` guard to ensure correctness.
- The application of migrations is done as close to `Main` as possible
(literally in `Main` for Quine, but in the cluster join message handler
for thatDot Streaming Graph). This limits the information available to
the migration tools, as nothing but the namespaces have been restored by
the time migrations run. A completion of this system will require
between "several" and "many" lifecycle hooks that migrations can
implement, rather than a single `run`

## Type of change
- [x] New feature (non-breaking change which adds functionality)

# How Has This Been Tested?
For each of the following configurations, product version 1.6.2 was
loaded, a standing query was registered, a node was created, the product
was closed, then `HEAD` was loaded. The output was confirmed sane in
each case. Note that not all combinations are represented, but I believe
these provide sufficient coverage. To help explain, I've annotated each
test with why it's included

In the case of ClickHouse and Cassandra, the instructions presented by
the product were followed and the product was confirmed to operate
normally

- [x] Quine / RocksDB (most common OSS setup)
- [x] Quine / Cassandra (likely setup for people on "give us money"
path, which we of course would like to encourage)
- [x] ~Quine / MapDB (representative of fallback / "other not special
cased but not migratable persistor" case)~ N/A, mapDB is broken QU-1920
- [x] Quine / Empty (representative of "no migration needed" case)
- [x] thatDot Streaming Graph / Cassandra (most common enterprise setup)
- [x] thatDot Streaming Graph / ClickHouse (the other migratable
enterprise persistor)
- [x] thatDot Streaming Graph / RocksDB (uncommon non-migratable
enterprise setup)
- [x] ~thatDot Streaming Graph / MapDB (representative of fallback case
given obfuscation)~ N/A, mapDB is broken QU-1920

Additionally:
- [x] A 2-member thatDot Streaming Graph cluster was started with
version 1.6.4, one member was killed, and a replacement member was
started up with 1.7.0 to confirm that the 1.7.0 host prevented the
cluster from returning to Operating in an inconsistent state. This is
meant to smoke-test an analogy to the k8s rolling redeploy

# Checklist:
- [x] I have performed a self-review of my code
- [x] I have verified my code doesn't add an implementation for
something that already exists
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] New and existing unit tests pass locally with my changes

GitOrigin-RevId: 19a948680dd69b75a9a1d3c2a1645675d5a2b42c